backend/requirements-voice-cloning.txt
# Voice Cloning Dependencies
coqui-tts>=0.14.0
torch>=1.13.0
torchaudio>=0.13.0
numpy>=1.21.0
soundfile>=0.12.0
librosa>=0.9.0
scipy>=1.9.0
matplotlib>=3.5.0

# Additional dependencies for voice cloning
phonemizer>=3.0.0
espeak-ng>=1.51
ffmpeg-python>=0.2.0

# Core dependencies (if not already installed)
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0 